# avm2: share binding skip optimization with `call_method`