This PR shares the optimization from `call_property` which skips binding methods that don't need receivers with `call_method`. Currently, for example, Chess Demons allocates almost 1GB of unnecessary function objects during loading, this patch fixes that:
```
Command                            USS 
./ruffle_desktop_opt cd.swf       1.5G 
./ruffle_desktop_no_opt cd.swf    2.5G
```
A quick `perf stat` run shows a small decrease in loading time and page faults cut in half:
```
Performance counter stats for './ruffle_desktop_opt cd.swf':

        20,082.41 msec task-clock:u                     #    0.582 CPUs utilized             
                0      context-switches:u               #    0.000 /sec                      
                0      cpu-migrations:u                 #    0.000 /sec                      
          376,719      page-faults:u                    #   18.759 K/sec                     
   94,845,384,958      cycles:u                         #    4.723 GHz                       
  182,895,351,746      instructions:u                   #    1.93  insn per cycle            
   29,683,116,696      branches:u                       #    1.478 G/sec                     
      199,608,906      branch-misses:u                  #    0.67% of all branches           

     34.511689417 seconds time elapsed

     18.827233000 seconds user
      1.218733000 seconds sys
```
```
Performance counter stats for './ruffle_desktop_no_opt cd.swf':

        21,512.53 msec task-clock:u                     #    0.592 CPUs utilized             
                0      context-switches:u               #    0.000 /sec                      
                0      cpu-migrations:u                 #    0.000 /sec                      
          634,737      page-faults:u                    #   29.505 K/sec                     
  100,060,811,478      cycles:u                         #    4.651 GHz                       
  187,065,447,138      instructions:u                   #    1.87  insn per cycle            
   30,586,698,113      branches:u                       #    1.422 G/sec                     
      189,874,657      branch-misses:u                  #    0.62% of all branches           

     36.345462896 seconds time elapsed

     19.933122000 seconds user
      1.542323000 seconds sys
```
Thanks to @adrian17 for figuring out the problem and suggesting this fix!